Move DNS method CheckCRCLocalDNSReachableFromHost with others of the same type
